[devil] Allow instantiation of AdbWrapper with USB bus identifer.
Adb allows selection of a device using both a serial number and the
device's USB bus identifier. This change makes it possible to create an
AdbWrapper instance using a USB bus identifier in addition to a serial
number. This option is required when multiple connected devices present
the same serial number.